refactor: 패키지 구조 정리 및 도메인 이벤트 기반 의존성 분리
