+++ src/main/java/io/wisoft/ignoa_api/auction/listener/AuctionExpiredListener.java
@@ -1,6 +1,6 @@
 package io.wisoft.ignoa_api.auction.listener;
 
-import io.wisoft.ignoa_api.auction.service.AuctionCloseManager;
+import io.wisoft.ignoa_api.auction.service.AuctionCloseService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.connection.Message;
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class AuctionExpiredListener implements MessageListener {
 
-    private final AuctionCloseManager auctionCloseManager;
+    private final AuctionCloseService auctionCloseService;
     private static final String AUCTION_KEY_PREFIX = "auction:";
 
     @Override
@@ -25,7 +25,7 @@ public class AuctionExpiredListener implements MessageListener {
             }
 
             long itemId = Long.parseLong(expiredKey.substring(AUCTION_KEY_PREFIX.length()));
-            auctionCloseManager.closeAuction(itemId);
+            auctionCloseService.closeAuction(itemId);
         } catch (Exception e) {
             log.error("경매 마감 처리 실패 expiredKey={}", message.toString(), e);
         }
