+++ src/main/java/io/wisoft/ignoa_api/auction/listener/AuctionRegistrationListener.java
@@ -2,8 +2,10 @@ package io.wisoft.ignoa_api.auction.listener;
 
 import io.wisoft.ignoa_api.auction.event.AuctionRegisteredEvent;
 import io.wisoft.ignoa_api.auction.service.AuctionRedisService;
+import io.wisoft.ignoa_api.item.event.ItemDeletedEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
 
@@ -16,5 +18,10 @@ public class AuctionRegistrationListener {
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     public void onAuctionRegistered(AuctionRegisteredEvent event) {
         auctionRedisService.registerTtl(event.itemId(), event.endAt());
+    }
+
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    public void onItemDeleted(ItemDeletedEvent event) {
+        auctionRedisService.deleteTtl(event.itemId());
     }
 }
