+++ src/main/java/io/wisoft/ignoa_api/bid/controller/BidController.java
@@ -5,8 +5,8 @@ import io.wisoft.ignoa_api.bid.dto.request.BidCreateRequest;
 import io.wisoft.ignoa_api.bid.dto.response.BidSummary;
 import io.wisoft.ignoa_api.bid.dto.response.BidResponse;
 import io.wisoft.ignoa_api.bid.service.BidService;
-import io.wisoft.ignoa_api.global.dto.ApiResponse;
-import io.wisoft.ignoa_api.global.dto.SliceResponse;
+import io.wisoft.ignoa_api.global.common.ApiResponse;
+import io.wisoft.ignoa_api.global.common.SliceResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
