+++ src/main/java/io/wisoft/ignoa_api/bid/entity/Bid.java
@@ -2,7 +2,7 @@ package io.wisoft.ignoa_api.bid.entity;
 
 import io.wisoft.ignoa_api.item.entity.Item;
 import io.wisoft.ignoa_api.user.entity.User;
-import io.wisoft.ignoa_api.global.entity.BaseEntity;
+import io.wisoft.ignoa_api.global.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
