+++ src/main/java/io/wisoft/ignoa_api/global/common/SliceResponse.java
@@ -1,4 +1,4 @@
-package io.wisoft.ignoa_api.global.dto;
+package io.wisoft.ignoa_api.global.common;
 
 import java.util.List;
