+++ src/main/java/io/wisoft/ignoa_api/global/config/SecurityConfig.java
@@ -32,7 +32,6 @@ public class SecurityConfig {
                         .requestMatchers("/api/auth/**").permitAll()
                         .requestMatchers("/api/users/email/duplicate").permitAll()
                         .requestMatchers("/api/users/nickname/duplicate").permitAll()
-                        .requestMatchers("/swagger-ui/**", "/v3/api-docs/**").permitAll()
                         .requestMatchers("/ws/**").permitAll()
                         .anyRequest().authenticated()
                 ).addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class)
