--- src/main/java/io/wisoft/ignoa_api/global/config/SwaggerConfig.java
@@ -1,27 +1,0 @@
-package io.wisoft.ignoa_api.global.config;
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.security.SecurityRequirement;
-import io.swagger.v3.oas.models.security.SecurityScheme;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI openAPI() {
-        SecurityScheme securityScheme = new SecurityScheme()
-                .type(SecurityScheme.Type.HTTP)
-                .scheme("bearer")
-                .bearerFormat("JWT")
-                .in(SecurityScheme.In.HEADER)
-                .name("Authorization");
-
-        return new OpenAPI()
-                .components(new Components()
-                        .addSecuritySchemes("bearerAuth", securityScheme))
-                .addSecurityItem(new SecurityRequirement().addList("bearerAuth"));
-    }
-}
