--- src/main/java/io/wisoft/ignoa_api/item/controller/ItemSseController.java
@@ -1,24 +1,0 @@
-package io.wisoft.ignoa_api.item.controller;
-
-import io.wisoft.ignoa_api.item.sse.SseEmitterManager;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
-
-
-@RestController
-@RequestMapping("/api/items/{itemId}")
-@RequiredArgsConstructor
-public class ItemSseController {
-
-    private final SseEmitterManager sseEmitterManager;
-
-    @GetMapping(value = "/subscribe", produces = MediaType.TEXT_EVENT_STREAM_VALUE)
-    public SseEmitter subscribe(@PathVariable Long itemId) {
-        return sseEmitterManager.subscribe(itemId);
-    }
-}
