+++ src/main/java/io/wisoft/ignoa_api/item/event/ItemDeletedEvent.java
@@ -1,0 +1,6 @@
+package io.wisoft.ignoa_api.item.event;
+
+public record ItemDeletedEvent(
+        Long itemId
+) {
+}
