--- src/main/java/io/wisoft/ignoa_api/storage/dto/response/PresignedUrlResponse.java
@@ -1,7 +1,0 @@
-package io.wisoft.ignoa_api.storage.dto.response;
-
-public record PresignedUrlResponse(
-        String presignedUrl,
-        String objectUrl
-) {
-}
