+++ src/main/java/io/wisoft/ignoa_api/user/entity/User.java
@@ -1,6 +1,6 @@
 package io.wisoft.ignoa_api.user.entity;
 
-import io.wisoft.ignoa_api.global.entity.BaseEntity;
+import io.wisoft.ignoa_api.global.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
