+++ src/main/resources/application.yaml
@@ -1,6 +1,12 @@
 server:
   port: 38080
 
+logging:
+  pattern:
+    console: "%clr(%d{HH:mm:ss}){faint} %clr([%15.15thread]){magenta} %clr(%-5level){highlight} %clr(%logger{36}){cyan} - %msg%n"
+  level:
+    io.wisoft.prepair: DEBUG
+
 spring:
   application:
     name: Ignoa API
@@ -40,7 +46,6 @@ cloud:
       access-key: ${S3_ACCESS_KEY}
       secret-key: ${S3_SECRET_KEY}
       region: ${S3_REGION}
-      presigned-url-expiration: ${S3_PRESIGNED_EXPIRATION}
 
 jwt:
   secret: ${JWT_SECRET}
